+++ streamn_scoreboard_plugin/__init__.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-__version__ = "0.5.4"
+__version__ = "0.6.0"
 
 from streamn_scoreboard_plugin.plugin import ScoreboardPlugin